requirements.txt
# Python 3.8
dm-logger~=0.6.2
python-dotenv>=1.0.0
mysql-connector-python>=9.0.0, <10.0.0
